# api/docs/swagger.yaml
definitions:
  models.BookPrimeryKey:
    properties:
      id:
        type: string
    type: object
  models.Category:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  models.CreateBook:
    properties:
      created_at:
        type: string
      description:
        type: string
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
  models.CreateCategory:
    properties:
      name:
        type: string
    type: object
  models.Empty:
    type: object
  models.GetListBookRequest:
    properties:
      limit:
        type: integer
      offset:
        type: integer
    type: object
  models.GetListCategoryResponse:
    properties:
      Categorys:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      count:
        type: integer
    type: object
  models.PutBook:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  models.UpdateCategorySwag:
    properties:
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /book:
    get:
      consumes:
      - application/json
      description: Get List Book
      operationId: Get_List_Book
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: GetListBookBody
          schema:
            $ref: '#/definitions/models.GetListBookRequest'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get List Book
      tags:
      - Book
    post:
      consumes:
      - application/json
      description: Create Book
      operationId: create_Book
      parameters:
      - description: CreateBookRequestBody
        in: body
        name: Book
        required: true
        schema:
          $ref: '#/definitions/models.PutBook'
      produces:
      - application/json
      responses:
        "201":
          description: GetBookryBody
          schema:
            $ref: '#/definitions/models.CreateBook'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Create Book
      tags:
      - Book
  /book/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Book
      operationId: Delete_Book
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: DeleteBookyBody
          schema:
            $ref: '#/definitions/models.Empty'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Delete Book
      tags:
      - Book
    get:
      consumes:
      - application/json
      description: GetByID Book
      operationId: Get_By_IDBook
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: GetByIDBookBody
          schema:
            $ref: '#/definitions/models.BookPrimeryKey'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: GetByID Book
      tags:
      - Book
    put:
      consumes:
      - application/json
      description: Update_Book
      operationId: Update_Book
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateBookResquestBody
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.PutBook'
      produces:
      - application/json
      responses:
        "202":
          description: UpdateBookBody
          schema:
            $ref: '#/definitions/models.PutBook'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Update_Book
      tags:
      - Book
  /category:
    get:
      consumes:
      - application/json
      description: Get List Category
      operationId: get_list_category
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: GetCategoryListBody
          schema:
            $ref: '#/definitions/models.GetListCategoryResponse'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get List Category
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: Create Category
      operationId: create_category
      parameters:
      - description: CreateCategoryRequestBody
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.CreateCategory'
      produces:
      - application/json
      responses:
        "201":
          description: GetCategoryBody
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Create Category
      tags:
      - Category
  /category/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Category
      operationId: delete_category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: DeleteCategoryBody
          schema:
            $ref: '#/definitions/models.Empty'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Delete Category
      tags:
      - Category
    get:
      consumes:
      - application/json
      description: Get By ID Category
      operationId: get_by_id_category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetCategoryBody
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get By ID Category
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: Update Category
      operationId: update_category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateCategoryRequestBody
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCategorySwag'
      produces:
      - application/json
      responses:
        "202":
          description: UpdateCategoryBody
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Update Category
      tags:
      - Category
swagger: "2.0"